Replace isinstance(x, unicode) with not isinstance(x, str)
Qin Zhang wanted to replace isinstance(x, unicode) with isinstance(x, bytes))
but this breaks Python2 without a need to do so.

As these checks are at places where we want to convert to "str",
we can just use this instead:

   if not isinstance(x, str):

And then convert to str. Its trivial and works for both versions equally.

Co-Authored-by: Qin Zhang （张琴） <qin.zhang@citrix.com>
Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>